pubsub-python-sample/pub-python-pubsub-test/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==1.1.0